Models and Mappers
Instead of passing the raw data to my application, I'm passing it a model
which is a thin wrapper around the data, with getters for the properties
my application is expecting. This will make guarding against data inconsistencies
much easier. It also means any manipulations on the data, or computed properties,
can be executed on demand instead of on application loading.